% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{agglomerate_data}
\alias{agglomerate_data}
\title{Agglomerate ABRP 16S data to desired taxonomic level}
\usage{
agglomerate_data(tax_level = "genus", replicates = TRUE)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{replicates}{include replicates}
}
\value{
phyloseq object
}
\description{
Agglomerate ABRP 16S data to desired taxonomic level
}
\examples{
data <- agglomerate_data(tax_level="genus", replicates=TRUE)
}
