% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{build_design_matrix}
\alias{build_design_matrix}
\title{Build the design matrix}
\usage{
build_design_matrix(metadata, metadata.diet = NULL)
}
\arguments{
\item{metadata}{metadata data.frame associated with the data phyloseq object}

\item{metadata.diet}{metadata data.frame parsed from separate diet and climate data}
}
\value{
matrix with observation time (days), diet PCs, and climate (rainfall and max temp) covariates
}
\description{
Build the design matrix
}
