% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{build_design_matrix_predict}
\alias{build_design_matrix_predict}
\title{Build the design matrix for prediction by interpolating within the}
\usage{
build_design_matrix_predict(X)
}
\arguments{
\item{X}{existing design matrix}
}
\value{
matrix with observation time (days), diet PCs, and climate (rainfall and max temp) covariates interpolated from
earliest to latest observation date
}
\description{
Build the design matrix for prediction by interpolating within the
}
