% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calc_autocorrelation}
\alias{calc_autocorrelation}
\title{Calculate lagged autocorrelation}
\usage{
calc_autocorrelation(data, lag_units = "weeks", lag_max = 52,
  use_lr = "ilr", alr_ref = NULL, resample = FALSE,
  resample_rate = 0.2)
}
\arguments{
\item{data}{a phyloseq object}

\item{lag_units}{units of autocorrelation lag ("weeks", "months", "seasons")}

\item{lag_max}{maximum lagged difference to evaluate}

\item{use_lr}{log ratio representation ("ilr", "alr", "clr")}

\item{alr_ref}{if using the additive log ratio (ALR), which reference taxon index to use}

\item{resample}{use resampling for uncertainty estimation}

\item{resample_rate}{percent of samples to randomly subsample if resampling}
}
\value{
data.frame of (bounded, if resampled) autocorrelation, indexed by lag in lag_units
}
\description{
Calculate lagged autocorrelation
}
\examples{
lagged_ac <- calc_autocorrelation(data, lag_units="weeks", lag_max=52, use_lr="ilr", alr_ref=NULL, resample=FALSE, resample_rate=0.2)
}
