% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{calc_microbe_universality_score}
\alias{calc_microbe_universality_score}
\title{Calculate a score (0.0 to 1.0) for the "universality" of given microbe}
\usage{
calc_microbe_universality_score(tax_level = "ASV", Sigmas = NULL)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{Sigmas}{optional list (indexed by host short name) of MAP estimates of microbial covariance; if not provided, this will be loaded}
}
\value{
named list of per-microbe scores
}
\description{
Calculate a score (0.0 to 1.0) for the "universality" of given microbe
}
\details{
These scores do not currently incorporate posterior uncertainty. Returns results for correlations in CLR.
}
\examples{
Sigmas <- load_MAP_estimates(tax_level="ASV", logratio="clr")
scores <- calc_microbe_universality_score(tax_level="ASV", Sigmas=Sigmas)
reorder <- order(unlist(scores), decreasing=TRUE)
scores <- unlist(scores)[reorder]
}
