% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{calc_posterior_distance_row}
\alias{calc_posterior_distance_row}
\title{Calculate Riemannian distances between one sample and all others; this is to allow extremely
large distance matrices to be calculated in blocks by independent parallel jobs on the
cluster}
\usage{
calc_posterior_distance_row(tax_level = "ASV", sample_idx = 1)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{sample_idx}{the posterior sample to be considered against all others}
}
\description{
Calculate Riemannian distances between one sample and all others; this is to allow extremely
large distance matrices to be calculated in blocks by independent parallel jobs on the
cluster
}
\details{
Output is saved to a file.
}
\examples{
calc_posterior_distances_row(tax_level="ASV", sample_idx=1)
}
