% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{calc_universality_score}
\alias{calc_universality_score}
\title{Calculate a score (0.0 to 1.0) for the "universality" of given microbial interaction}
\usage{
calc_universality_score(x)
}
\arguments{
\item{x}{vector of correlations between a pair of CLR microbes across individuals}
}
\value{
score
}
\description{
Calculate a score (0.0 to 1.0) for the "universality" of given microbial interaction
}
\details{
An element of x should be the (MAP) correlation between a pair of CLR microbes.
The length of x is assumed equal to the number of hosts.
}
\examples{
interaction_matrix <- get_universal_interactions(tax_level="ASV")
score <- calc_universality_score(interaction_matrix[,1])
}
