% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R
\name{calculate_emd}
\alias{calculate_emd}
\title{Calculate Earth Mover's Distance between two 2D summaries}
\usage{
calculate_emd(heatmap1, heatmap2)
}
\arguments{
\item{heatmap1}{first matrix of correlations (rows are hosts, columns are taxon-taxon pairs)}

\item{heatmap2}{second matrix of correlations (rows are hosts, columns are taxon-taxon pairs)}
}
\description{
Calculate Earth Mover's Distance between two 2D summaries
}
