% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{check_output_dir}
\alias{check_output_dir}
\title{Create output directory/ies that may not already exist}
\usage{
check_output_dir(path_to_dir)
}
\arguments{
\item{path_to_dir}{a character vector of subdirectories to create (if not already in existence)}
}
\value{
the path as a joined string
}
\description{
Create output directory/ies that may not already exist
}
\examples{
check_output_dir(list("output", "plots"))
}
