% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{default_ALR_prior}
\alias{default_ALR_prior}
\title{Set up a basic ALR prior}
\usage{
default_ALR_prior(D, log_var_scale = 1)
}
\arguments{
\item{D}{number of features including reference (where the ALR will represent D-1)}

\item{log_var_scale}{scale of the log variance}
}
\value{
list containing inverse Wishart parameters degrees of freedom and scale matrix
}
\description{
Set up a basic ALR prior
}
\examples{
params <- default_ALR_prior(D=100, log_var_scale=1)
Sigma <- matrixsampling::rinvwishart(1, params$upsilon, params$Xi)[,,1]
}
