% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{embed_posteriors}
\alias{embed_posteriors}
\title{Embeds posterior samples using MDS and a pre-calculated distance matrix}
\usage{
embed_posteriors(tax_level = "genus", which_measure = "Sigma", MAP = FALSE)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{which_measure}{estimated object to embed, either Lambda or Sigma}

\item{MAP}{use MAP estimate model output instead of full posterior output}
}
\description{
Embeds posterior samples using MDS and a pre-calculated distance matrix
}
\details{
Distance matrix between posterior samples must be present in designated output directory
}
\examples{
embed_posteriors(tax_level="genus", which_measure="Sigma", MAP=FALSE)
}
