% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{filter_data}
\alias{filter_data}
\title{Filter out low abundance taxa}
\usage{
filter_data(data, tax_level = NULL, count_threshold = 5,
  sample_threshold = 0.2)
}
\arguments{
\item{data}{a phyloseq object}

\item{tax_level}{taxonomic level at which data have been agglomerated}

\item{count_threshold}{minimum count at which a taxon must be observed}

\item{sample_threshold}{minimum proportion of samples within each host at which a taxon must be observed at the level specified by count_threshold}
}
\value{
phyloseq object
}
\description{
Filter out low abundance taxa
}
\details{
Together count_threshold and sample_threshold specify a minimum representation for a taxon.
Taxa below these observed thresholds will be grouped together into a category "Collapsed".
}
\examples{
filtered_data <- filter_data(data, count_threshold=5, sample_threshold=0.2)
}
