% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualizing_distance.R
\name{filter_taxa_2D_summary}
\alias{filter_taxa_2D_summary}
\title{Given a count table, label taxa with at least 0.1% relative abundance}
\usage{
filter_taxa_2D_summary(counts)
}
\arguments{
\item{counts}{count table (taxa as rows, samples as columns)}
}
\value{
number of unique combinations of features
}
\description{
Given a count table, label taxa with at least 0.1% relative abundance
}
