% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_DLM}
\alias{fit_DLM}
\title{Fit a dynamic linear model to a single host series using labraduck}
\usage{
fit_DLM(
  data,
  host,
  taxa_covariance,
  var_scale = 1,
  tax_level = "ASV",
  alr_ref = NULL,
  n_samples = 100,
  MAP = FALSE
)
}
\arguments{
\item{data}{a phyloseq object}

\item{host}{host short name (e.g. ACA)}

\item{taxa_covariance}{list of prior covariance parameters over taxa}

\item{var_scale}{combined scale of the total variance components Sigma and Gamma (default 1)}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{alr_ref}{index of reference ALR coordinate}

\item{n_samples}{number of posterior samples to draw}

\item{MAP}{compute MAP estimate only (as single posterior sample)}
}
\description{
Fit a dynamic linear model to a single host series using labraduck
}
\details{
Fitted model and metadata saved to designated model output directory.
}
\examples{
tax_level <- "ASV"
data <- load_data(tax_level = tax_level)
params <- formalize_parameters(data)
taxa_covariance <- get_Xi(phyloseq::ntaxa(data), total_variance = 1)
fit_DLM(data, host = "GAB", taxa_covariance = taxa_covariance, tax_level = tax_level, alr_ref = params$alr_ref, MAP = TRUE)
}
