% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_GP}
\alias{fit_GP}
\title{Fit a Gaussian process to a single host series using basset}
\usage{
fit_GP(data, host, taxa_covariance, sample_covariance, tax_level = "ASV",
  alr_ref = NULL, n_samples = 100, MAP = FALSE,
  holdout_proportion = 0, return_model = FALSE, scramble = FALSE)
}
\arguments{
\item{data}{a phyloseq object}

\item{host}{host short name (e.g. ACA)}

\item{taxa_covariance}{list of prior covariance parameters over taxa}

\item{sample_covariance}{kernel function}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{alr_ref}{index of reference ALR coordinate}

\item{n_samples}{number of posterior samples to draw}

\item{MAP}{compute MAP estimate only (as single posterior sample)}

\item{holdout_proportion}{if non-zero, proportion of host's sample to use as a test set}

\item{return_model}{if TRUE, returns the fitted model instead of saving it}

\item{scramble}{if TRUE, host samples are scrambled in time; this serve the diagnostic purpose of 
identifying whether some of the correlation we see between individual hosts' dynamics
are the result of systematic differences in abundances (etc.)}

\item{rho}{bandwidth for SE kernel}
}
\description{
Fit a Gaussian process to a single host series using basset
}
\details{
Fitted model and metadata saved to designated model output directory.
}
\examples{
tax_level <- "ASV"
data <- load_data(tax_level = tax_level)
params <- formalize_parameters(data)
sample_covariance <- get_Gamma(kernel_scale = 2, proportions = c(1, 0, 0), min_correlation = 0.1, days_to_baseline = 90)
taxa_covariance <- get_Xi(phyloseq::ntaxa(data), total_variance = 1)
fit_GP(data, host = "GAB", taxa_covariance = taxa_covariance, sample_covariance = sample_covariance, tax_level = tax_level, alr_ref = params$alr_ref, MAP = TRUE)
}
