% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fix_MAP_dims}
\alias{fix_MAP_dims}
\title{Trivially expands the dimensions of the posterior estimates objects in a MAP estimate basset result}
\usage{
fix_MAP_dims(fit)
}
\arguments{
\item{fit}{a bassetfit object}
}
\description{
Trivially expands the dimensions of the posterior estimates objects in a MAP estimate basset result
}
\examples{
fit <- fix_MAP_dims(fit)
}
