% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fold_error}
\alias{fold_error}
\title{Utility function to calculate fold change (as error)}
\usage{
fold_error(y1, y2, pc = 0.5)
}
\arguments{
\item{y1}{count vector}

\item{y2}{count vector}

\item{pc}{optional pseudocount parameter}
}
\value{
average fold difference between vectors
}
\description{
Utility function to calculate fold change (as error)
}
