% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{formalize_parameters}
\alias{formalize_parameters}
\title{Generate empirical estimates for variance components to use in the Gaussian process}
\usage{
formalize_parameters(data)
}
\arguments{
\item{data}{a phyloseq object}
}
\value{
list containing median log abundance taxon and mean empirical total variance
across hosts using this ALR representation
}
\description{
Generate empirical estimates for variance components to use in the Gaussian process
}
