% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R
\name{generate_highcontrast_palette}
\alias{generate_highcontrast_palette}
\title{Generate a random high-contrast palette (for bar plotting)}
\usage{
generate_highcontrast_palette(S)
}
\arguments{
\item{S}{desired palette size}
}
\value{
vector of hex color strings
}
\description{
Generate a random high-contrast palette (for bar plotting)
}
