% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{get_all_vs_one_correlations}
\alias{get_all_vs_one_correlations}
\title{Get correlations between one microbe and all others at designated taxonomic level}
\usage{
get_all_vs_one_correlations(
  taxon_idx,
  tax_level = "ASV",
  logratio = "alr",
  Sigmas = NULL
)
}
\arguments{
\item{taxon_idx}{the logratio coordinate to render correlations against}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{logratio}{logratio representation to use (e.g. "alr", "ilr", "clr")}

\item{Sigmas}{optional list (indexed by host short name) of MAP estimates of microbial covariance; if not provided, this will be loaded}
}
\value{
correlation matrix of dimensions {number of hosts} x {number of logratio taxa - 1}
}
\description{
Get correlations between one microbe and all others at designated taxonomic level
}
\examples{
interactions <- get_all_vs_one_correlations(taxon_idx = 1, tax_level = "ASV", logratio = "alr")
}
