% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{get_deepest_assignment}
\alias{get_deepest_assignment}
\title{Get highest taxonomic level to which a microbe is resolved (i.e. not NA)}
\usage{
get_deepest_assignment(taxonomy, deepest_tax_level = "genus")
}
\arguments{
\item{taxonomy}{a named list of taxonomic assignment for a designated sequence variant}

\item{deepest_tax_level}{the finest resolution taxonomy level to return (if possible)}
}
\value{
readable label of the form "phylum Tenericutes"
}
\description{
Get highest taxonomic level to which a microbe is resolved (i.e. not NA)
}
\examples{
data <- load_data(tax_level="ASV")
alr_ref <- formalize_parameters(data)$alr_ref
taxonomy <- get_taxonomy(data, alr_ref)
# get deepest resolve level for the first sequence variant
label <- get_deepest_assignment(taxonomy[1,], deepest_tax_level="species")
}
