% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{get_fitted_model_list}
\alias{get_fitted_model_list}
\title{Pulls a list of fitted basset models from the designated output directory}
\usage{
get_fitted_model_list(tax_level = "ASV", MAP = FALSE)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{MAP}{use MAP estimate model output instead of full posterior output}
}
\description{
Pulls a list of fitted basset models from the designated output directory
}
\examples{
model_list <- get_fitted_model_list(tax_level="ASV", MAP=FALSE)
}
