% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{get_group_labels}
\alias{get_group_labels}
\title{Get primary social group for all hosts in data set}
\usage{
get_group_labels(data)
}
\value{
named list
}
\description{
Get primary social group for all hosts in data set
}
\details{
Output is a data.frame of fitness annotations sorted by host short name
}
\examples{
group_labels <- get_group_labels(data)
}
