% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{get_other_labels}
\alias{get_other_labels}
\title{Get other annotations (e.g. sample number) for hosts in data set}
\usage{
get_other_labels(data, tax_level = "genus", annotation = "group", MAP = FALSE)
}
\arguments{
\item{data}{a phyloseq object}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{annotation}{label to assign (e.g. individual)}

\item{MAP}{use MAP estimate model output instead of full posterior output}
}
\value{
data.frame
}
\description{
Get other annotations (e.g. sample number) for hosts in data set
}
\details{
Output is the centroids data.frame appended with the desired annotation. Available annotations are:
group, matgroup, samplenumber, sampledensity, mom, dad, momrank, drought, largegroup, 
momdied, competingsib, earlyadversity, birthrate_all, birthrate_surviving, alphadiv,
wetproportion
}
\examples{
labelled_centoids <- get_other_labels(data, tax_level="genus", annotation="group", MAP=FALSE)
}
