% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{get_pairwise_proportionalities}
\alias{get_pairwise_proportionalities}
\title{Get all pairwise (MAP) proportionalities between microbes at designated taxonomic level}
\usage{
get_pairwise_proportionalities(
  tax_level = "ASV",
  logratio = "alr",
  Sigmas = NULL
)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{logratio}{logratio representation to use (e.g. "alr", "ilr", "clr")}

\item{Sigmas}{optional list (indexed by host short name) of MAP estimates of microbial covariance; if not provided, this will be loaded}

\item{taxon_idx}{the logratio coordinate to render correlations against}
}
\value{
correlation matrix of dimensions {number of hosts} x {number of unique interactions between logratio taxa}
}
\description{
Get all pairwise (MAP) proportionalities between microbes at designated taxonomic level
}
\examples{
interactions <- get_pairwise_proportionalities(tax_level = "ASV", logratio = "alr")
}
