% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{get_Rdist_pair}
\alias{get_Rdist_pair}
\title{Calculate pairwise distances (slowly) by calling Riemann_dist_pair
This is just a convenience function}
\usage{
get_Rdist_pair(A, B)
}
\arguments{
\item{A}{covariance/correlation matrix 1}

\item{B}{covariance/correlation matrix 2}
}
\value{
(scalar) Riemannian distance between matrices
}
\description{
Calculate pairwise distances (slowly) by calling Riemann_dist_pair
This is just a convenience function
}
