% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Get taxonomy, reordered to reflect altered ALR reference}
\usage{
get_taxonomy(data, alr_ref)
}
\arguments{
\item{data}{a phyloseq object}

\item{alr_ref}{index of ALR reference taxon}
}
\value{
data.frame of re-ordered taxonomy
}
\description{
Get taxonomy, reordered to reflect altered ALR reference
}
\examples{
tax <- get_taxonomy(data, alr_ref=67)
}
