% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{get_universal_interactions}
\alias{get_universal_interactions}
\title{Identify and plot "universal" interactions at the designated taxonomic level}
\usage{
get_universal_interactions(
  tax_level = "ASV",
  show_plot = FALSE,
  order_by = "abundance"
)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{show_plot}{show() plot in addition to rendering it to a file}

\item{order_by}{ordering of bigraph components; options are "abundance" or "taxonomy"}
}
\value{
data.frame of microbial pairs (by index in taxonomy and label) and median correlation across hosts
}
\description{
Identify and plot "universal" interactions at the designated taxonomic level
}
\details{
Correlations are evaluated in the CLR. A threshold of 0.4 will identify pairs of 
taxa with median correlation of > 0.4 or < -0.4. Pairs are rendered as bigraphs.
}
\examples{
interaction_matrix <- get_universal_interactions(tax_level="ASV")
}
