% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualizing_distance.R
\name{load_ABRP_2D_summary}
\alias{load_ABRP_2D_summary}
\title{Parse correlation matrix for Amboseli data}
\usage{
load_ABRP_2D_summary(MAP = TRUE)
}
\arguments{
\item{heatmap}{vector or matrix (rows as hosts) of taxon-taxon correlations across hosts}
}
\description{
Parse correlation matrix for Amboseli data
}
\details{
Note: This version is using CLR correlation (not proportionality).
}
