% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{load_data}
\alias{load_data}
\title{Load ABRP 16S data and filter out low abundance taxa}
\usage{
load_data(
  tax_level = "genus",
  replicates = TRUE,
  host_sample_min = 40,
  count_threshold = 5,
  sample_threshold = 0.2
)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{replicates}{include replicates}

\item{host_sample_min}{minimum sample number for host inclusion in the filtered data set}

\item{count_threshold}{minimum count at which a taxon must be observed}

\item{sample_threshold}{minimum proportion of samples within each host at which a taxon must be observed at the level specified by count_threshold}
}
\value{
phyloseq object
}
\description{
Load ABRP 16S data and filter out low abundance taxa
}
\details{
Together count_threshold and sample_threshold specify a minimum representation for a taxon.
Taxa below these observed thresholds will be grouped together into a category "Other."
}
\examples{
data <- load_data(tax_level="genus", replicates=TRUE, host_sample_min=40, count_threshold=5, sample_threshold=0.2)
}
