% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{load_full_posteriors}
\alias{load_full_posteriors}
\title{Load full posterior estimates of microbial covariances from fitted models at the desired taxonomic level}
\usage{
load_full_posteriors(tax_level = "ASV", logratio = "alr")
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{logratio}{logratio representation to use (e.g. "alr", "ilr", "clr")}
}
\value{
list of full posterior covariance estimates indexed by host short name
}
\description{
Load full posterior estimates of microbial covariances from fitted models at the desired taxonomic level
}
\examples{
Sigmas <- load_full_posteriors(tax_level="ASV", logratio="alr")
}
