% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualizing_distance.R
\name{load_Johnson_2D_summary}
\alias{load_Johnson_2D_summary}
\title{Load data from Johnson et al. (2019)}
\usage{
load_Johnson_2D_summary(visualize = NULL, GP = FALSE)
}
\arguments{
\item{visualize}{optional parameter specifying the quick visualization of the data to use;
"ribbon" fits and visualizes the fido model, "bars" plots a stacked bar chart}

\item{GP}{if TRUE, fit a Gaussian process model as opposed to a dynamic linear model}
}
\value{
correlation matrix (hosts x taxon-taxon associations) or NULL
}
\description{
Load data from Johnson et al. (2019)
}
