% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{log_rmse}
\alias{log_rmse}
\title{Utility function to calculate RMSE}
\usage{
log_rmse(y1, y2, pc = 0.5)
}
\arguments{
\item{y1}{count vector}

\item{y2}{count vector}

\item{pc}{optional pseudocount parameter}
}
\value{
root mean squared error of log counts
}
\description{
Utility function to calculate RMSE
}
