% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{loo_cv}
\alias{loo_cv}
\title{Perform leave-one-out cross-validation for a given host and kernel configuration choice}
\usage{
loo_cv(data, host, kernels, tax_level = "ASV", iterations = NULL)
}
\arguments{
\item{data}{a phyloseq object}

\item{host}{host short name (e.g. ACA)}

\item{kernels}{composite kernel function}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{iterations}{number of CV iterations to perform; this should be <= n_samples for this host}
}
\description{
Perform leave-one-out cross-validation for a given host and kernel configuration choice
}
\details{
average error (currently log RMSE)
}
\examples{
tax_level <- "ASV"
data <- load_data(tax_level = tax_level)
kernels <- get_Gamma(kernel_scale = 2, proportions = c(1, 0, 0), days_to_baseline = 90)
loo_cv(data, host = "GAB", kernels <- kernels, tax_level = tax_level)
}
