% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{make_posterior_predictions}
\alias{make_posterior_predictions}
\title{Predict from fitted basset model}
\usage{
make_posterior_predictions(X, fit)
}
\arguments{
\item{X}{observations in days relative to baseline}

\item{fit}{bassetfit object}

\item{n_samples}{number of posterior samples to draw}
}
\value{
list of prediction input (days relative to baseline) and output (Eta samples)
}
\description{
Predict from fitted basset model
}
\details{
Predictions will interpolate from beginning to end of observations.
}
\examples{
make_posterior_predictions(X, fit, n_samples=100)
}
