% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualizing_distance.R
\name{n_pairwise_combos}
\alias{n_pairwise_combos}
\title{Utility function to calculate the number of unique combinations of P features}
\usage{
n_pairwise_combos(P)
}
\arguments{
\item{P}{feature number}
}
\value{
number of unique combinations of features
}
\description{
Utility function to calculate the number of unique combinations of P features
}
