% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{plot_axes}
\alias{plot_axes}
\title{Plot a pair of principal coordinates}
\usage{
plot_axes(coordinates = NULL, centroids, tax_level = "genus",
  axis1 = 1, axis2 = 2, annotation = "host", MAP = FALSE,
  show_plot = FALSE)
}
\arguments{
\item{coordinates}{data.frame of coordinates from posterior embedding}

\item{centroids}{data.frame of centroids from posterior embedding}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{axis1}{PCoA coordinate to display on x-axis}

\item{axis2}{PCoA coordinate to display on y-axis}

\item{annotation}{label to assign (e.g. individual)}

\item{MAP}{use MAP estimate model output instead of full posterior output}

\item{show_plot}{show() plot in addition to rendering it to a file}
}
\description{
Plot a pair of principal coordinates
}
\details{
Coordinates parameter is only necessary where we're plotting full host posteriors (i.e. where
annotation="host").
}
\examples{
tax_level <- "genus"
annotation <- "group"
MAP <- FALSE
centroids <- readRDS(file.path("output","plots",tax_level,"Sigma_ordination_centroids.rds"))
labelled_centroids <- get_other_labels(centroids=centroids, tax_level=tax_level, annotation=annotation, MAP=MAP)
plot_axes(coordinates=labelled_centroids, tax_level=tax_level, axis1=1, axis2=2, annotation=annotation, MAP=MAP)
}
