% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualizing_distance.R
\name{plot_bars}
\alias{plot_bars}
\title{Plot relative abundances as stacked bars}
\usage{
plot_bars(counts, palette = NULL)
}
\arguments{
\item{counts}{count table (taxa x samples)}

\item{palette}{optional color palette (as a vector of hex color strings)}
}
\description{
Plot relative abundances as stacked bars
}
