% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_bounded_autocorrelation}
\alias{plot_bounded_autocorrelation}
\title{Renders autocorrelation plot with uncertainty estimates}
\usage{
plot_bounded_autocorrelation(lagged_ac)
}
\arguments{
\item{lags}{data.frame of lags and estimated autocorrelation (output of calc_autocorrelation())}
}
\description{
Renders autocorrelation plot with uncertainty estimates
}
\examples{
lagged_ac <- calc_autocorrelation(data, lag_units="weeks", lag_max=52, use_lr="ilr", alr_ref=NULL, resample=TRUE, resample_rate=0.2)
plot_bounded_autocorrelation(lagged_ac)
}
