% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mean_autocorrelation}
\alias{plot_mean_autocorrelation}
\title{Renders autocorrelation plot (without uncertainty estimates)}
\usage{
plot_mean_autocorrelation(lagged_ac, show_plot = FALSE)
}
\arguments{
\item{show_plot}{show() plot in addition to rendering it to a file}

\item{lags}{data.frame of lags and estimated autocorrelation (output of calc_autocorrelation())}
}
\description{
Renders autocorrelation plot (without uncertainty estimates)
}
\examples{
lagged_ac <- calc_autocorrelation(data, lag_units="weeks", lag_max=52, use_lr="ilr", alr_ref=NULL, resample=FALSE, resample_rate=0.2)
plot_mean_autocorrelation(lagged_ac)
}
