% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{read_file}
\alias{read_file}
\title{Load file with existence check}
\usage{
read_file(filename)
}
\arguments{
\item{filename}{filename (and path)}
}
\value{
RDS file
}
\description{
Load file with existence check
}
\examples{
input_data <- read_file("input/filtered_genus_5_20.rds")
}
