% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Riemann_dist_pair}
\alias{Riemann_dist_pair}
\title{Calculate distances between positive-definite covariance matrices}
\usage{
Riemann_dist_pair(A, B)
}
\arguments{
\item{A}{matrix 1}

\item{B}{matrix 2}
}
\value{
distance
}
\description{
Calculate distances between positive-definite covariance matrices
}
