% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Riemann_dist_samples}
\alias{Riemann_dist_samples}
\title{Calculate distances between a matrix of positive-definite covariance matrices concatenated
horizontally}
\usage{
Riemann_dist_samples(samples, n_indiv, n_samples_per)
}
\arguments{
\item{samples}{posterior samples matrix, arranged by host and concatenated across columns}

\item{n_indiv}{number of hosts}

\item{n_samples_per}{number of samples per host}
}
\value{
distance
}
\description{
Calculate distances between a matrix of positive-definite covariance matrices concatenated
horizontally
}
\details{
Dimensions of the samples matrix are D rows x DN columns where D = no. of features
and N = no. of samples
}
