\name{ROL-package}
\alias{ROL-package}
\alias{ROL}
\docType{package}
\title{
\packageTitle{ROL}
}
\description{
\packageDescription{ROL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ROL}
\packageIndices{ROL}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ROL}

Maintainer: \packageMaintainer{ROL}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
