% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualizing_distance.R
\name{summarize_all_pairs_2D}
\alias{summarize_all_pairs_2D}
\title{Summarizes a matrix of correlations (rows are hosts, columns are taxon-taxon pairs) in 2D}
\usage{
summarize_all_pairs_2D(vectorized_Sigmas)
}
\arguments{
\item{heatmap}{vector or matrix (rows as hosts) of taxon-taxon correlations across hosts}
}
\description{
Summarizes a matrix of correlations (rows are hosts, columns are taxon-taxon pairs) in 2D
}
