% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLV_simulations.R
\name{summarize_pairs_2D}
\alias{summarize_pairs_2D}
\title{Helper function for hockeystick plot that calculates a 2D summary of a taxon-taxon association
across hosts}
\usage{
summarize_pairs_2D(heatmap)
}
\arguments{
\item{heatmap}{vector or matrix (rows as hosts) of taxon-taxon correlations across hosts}
}
\description{
Helper function for hockeystick plot that calculates a 2D summary of a taxon-taxon association
across hosts
}
