% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{assign_concise_taxonomy}
\alias{assign_concise_taxonomy}
\title{Calculate Riemannian distances between all posterior samples}
\usage{
assign_concise_taxonomy(tax_level = "genus", logratio = "alr",
  other_category = 0)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{logratio}{logratio representation to use (e.g. "alr", "ilr", "clr")}

\item{other_category}{"Other" taxonomic category of collapsed low abundance taxa (if available)}
}
\value{
readable list of taxonomic labels (e.g. of the form "CLR(phylum Tenericutes)")
}
\description{
Calculate Riemannian distances between all posterior samples
}
\details{
If other_category is provided, these taxa will be rendered as *LR(Other).
}
\examples{
tax_labels <- assign_concise_taxonomy(tax_level="genus", logratio="alr", other_category=0)
}
