% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{calc_posterior_distance_row}
\alias{calc_posterior_distance_row}
\title{Calculate Riemannian distances between one sample and all others; this is to allow extremely
large distance matrices to be calculated in blocks by independent parallel jobs on the
cluster}
\usage{
calc_posterior_distance_row(tax_level = "genus", sample_idx = 1)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{sample_range}{posterior sample indices to consider (e.g. c(1,10) calculates distances between all individual
posterior samples 1 through 10)}
}
\description{
Calculate Riemannian distances between one sample and all others; this is to allow extremely
large distance matrices to be calculated in blocks by independent parallel jobs on the
cluster
}
\details{
Distance matrix between posterior samples saved to designated output directory. Saves
a list containing distance matrix and labels of each row or column (by host).
}
\examples{
calc_posterior_distances_ranged(tax_level="genus", sample_range=c(1,10))
}
