% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{calc_posterior_distances}
\alias{calc_posterior_distances}
\title{Calculate Riemannian distances between all posterior samples}
\usage{
calc_posterior_distances(
  tax_level = "genus",
  which_measure = "Sigma",
  MAP = FALSE
)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{which_measure}{estimated object to embed, either Lambda or Sigma}

\item{MAP}{use MAP estimate model output instead of full posterior output}
}
\description{
Calculate Riemannian distances between all posterior samples
}
\details{
Distance matrix between posterior samples saved to designated output directory. Saves
a list containing distance matrix and labels of each row or column (by host).
}
\examples{
calc_posterior_distances(tax_level="genus", which_measure="Sigma", MAP=FALSE)
}
