% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_GP}
\alias{fit_GP}
\title{Fit a Gaussian process to a single host series using basset}
\usage{
fit_GP(
  data,
  host,
  composite_kernel,
  rho,
  tax_level = "ASV",
  alr_ref = NULL,
  n_samples = 100,
  MAP = FALSE,
  holdout_proportion = 0
)
}
\arguments{
\item{data}{a phyloseq object}

\item{host}{host short name (e.g. ACA)}

\item{composite_kernel}{composite kernel function}

\item{rho}{bandwidth for SE kernel}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{alr_ref}{index of reference ALR coordinate}

\item{n_samples}{number of posterior samples to draw}

\item{MAP}{compute MAP estimate only (as single posterior sample)}

\item{holdout_proportion}{if non-zero, proportion of host's sample to use as a test set}
}
\description{
Fit a Gaussian process to a single host series using basset
}
\details{
Fitted model and metadata saved to designated model output directory.
}
\examples{
tax_level <- "ASV"
data <- load_data(tax_level = tax_level)
params <- formalize_parameters(data)
rho <- calc_se_decay(min_correlation = 0.1, days_to_baseline = 90)
composite_kernel <- get_Gamma(kernel_scale = 2, proportions = c(1, 0, 0), rho = rho)
fit_GP(data, host = "GAB", composite_kernel = composite_kernel, rho = rho, tax_level = tax_level, alr_ref = params$alr_ref, MAP = TRUE)
}
