% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_GP}
\alias{fit_GP}
\title{Fit a Gaussian process to a single host series using basset}
\usage{
fit_GP(data, host, tax_level = "ASV", SE_days_to_baseline = 90,
  date_lower_limit = NULL, date_upper_limit = NULL, alr_ref = NULL,
  n_samples = 100, MAP = FALSE, ...)
}
\arguments{
\item{data}{a phyloseq object}

\item{host}{host short name (e.g. ACA)}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{SE_days_to_baseline}{days at which squared exponential kernel decays to baseline correlation}

\item{date_lower_limit}{minimum date to consider (string format: YYYY-MM-DD)}

\item{date_upper_limit}{maximum date to consider (string format: YYYY-MM-DD)}

\item{alr_ref}{index of reference ALR coordinate}

\item{n_samples}{number of posterior samples to draw}

\item{MAP}{compute MAP estimate only (as single posterior sample)}
}
\description{
Fit a Gaussian process to a single host series using basset
}
\details{
Fitted model and metadata saved to designated model output directory.
}
\examples{
data <- load_data(tax_level="ASV")
fit_GP(data, host="ACA", tax_level="ASV", SE_sigma=1, PER_sigma=1, SE_days_to_baseline=90, MAP=FALSE)
}
