% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_GP}
\alias{fit_GP}
\title{Fit a Gaussian process to a single host series using basset}
\usage{
fit_GP(
  data,
  host,
  kernels,
  tax_level = "ASV",
  alr_ref = NULL,
  n_samples = 100,
  MAP = FALSE,
  loo = FALSE
)
}
\arguments{
\item{data}{a phyloseq object}

\item{host}{host short name (e.g. ACA)}

\item{kernels}{composite kernel function}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{alr_ref}{index of reference ALR coordinate}

\item{n_samples}{number of posterior samples to draw}

\item{MAP}{compute MAP estimate only (as single posterior sample)}

\item{loo}{if TRUE, one sample is left out, predicted, and RMSE calculated}
}
\description{
Fit a Gaussian process to a single host series using basset
}
\details{
Fitted model and metadata saved to designated model output directory.
}
\examples{
tax_level <- "ASV"
data <- load_data(tax_level = tax_level)
params <- formalize_parameters(data)
kernels <- get_Gamma(kernel_scale = 2, proportions = c(1, 0, 0), days_to_baseline = 90)
fit_GP(data, host = "GAB", tax_level = tax_level, kernels = kernels, alr_ref = params$alr_ref, MAP = TRUE)
}
