% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{formalize_parameters}
\alias{formalize_parameters}
\title{Generate empirical estimates for variance components to use in the Gaussian process by
estimating the noise between replicates}
\usage{
formalize_parameters(data)
}
\arguments{
\item{data}{a phyloseq object}
}
\value{
list containing variance recommended for SE and PER kernels
}
\description{
Generate empirical estimates for variance components to use in the Gaussian process by
estimating the noise between replicates
}
\details{
The SE and PER kernels are apportioned the empirical ALR variance, split 90/10 
between the kernels.
}
\examples{
params <- default_ALR_prior(D=100, log_var_scale=1)
Sigma <- matrixsampling::rinvwishart(1, params$upsilon, params$Xi)[,,1]
}
