% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{get_deepest_assignment}
\alias{get_deepest_assignment}
\title{Calculate Riemannian distances between all posterior samples}
\usage{
get_deepest_assignment(taxonomy, deepest_tax_level = "genus")
}
\arguments{
\item{taxonomy}{a named list of taxonomic assignment for a designated sequence variant}

\item{deepest_tax_level}{the finest resolution taxonomy level to return (if possible)}
}
\value{
readable label of the form "phylum Tenericutes"
}
\description{
Calculate Riemannian distances between all posterior samples
}
\examples{
data <- read_file(file.path("input","filtered_genus_5_20.rds"))
taxonomy <- tax_table(data)@.Data
# get deepest resolve level for the first sequence variant
label <- get_deepest_assignment(taxonomy[1,], deepest_tax_level="species")
}
