% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{get_Gamma}
\alias{get_Gamma}
\title{Define a kernel (function) over samples}
\usage{
get_Gamma(kernel_scale, proportions, days_to_baseline)
}
\arguments{
\item{kernel_scale}{total variance for the composite kernel}

\item{proportions}{proportion variance to attribute to each of 3 kernels (see details)}

\item{days_to_baseline}{days at which squared exponential kernel decays to baseline correlation of ~0.1}
}
\value{
kernel function
}
\description{
Define a kernel (function) over samples
}
\details{
Composite kernel is built from (1) squared exponential kernel (base autocorrelation component),
(2) seasonal kernel (periodic), and (3) diet and climate component (another squared exponential)
}
