% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{load_MAP_estimates}
\alias{load_MAP_estimates}
\title{Calculate Riemannian distances between all posterior samples}
\usage{
load_MAP_estimates(tax_level = "genus", logratio = "alr")
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{logratio}{logratio representation to use (e.g. "alr", "ilr", "clr")}
}
\value{
list of MAP covariance estimates indexed by host short name
}
\description{
Calculate Riemannian distances between all posterior samples
}
\examples{
Sigmas <- load_MAP_estimates(tax_level="genus", logratio="alr")
}
