% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_microbes.R
\name{plot_interaction_heatmap}
\alias{plot_interaction_heatmap}
\title{Plot heatmap over all pairwise (MAP) correlations between microbes at designated taxonomic level}
\usage{
plot_interaction_heatmap(tax_level = "ASV", logratio = "alr",
  Sigmas = NULL, cluster = TRUE, taxon_idx = NULL,
  show_plot = FALSE, return_matrix = FALSE)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{logratio}{logratio representation to use (e.g. "alr", "ilr", "clr")}

\item{Sigmas}{optional list (indexed by host short name) of MAP estimates of microbial covariance; if not provided, this will be loaded}

\item{cluster}{optional flag to hierarchically cluster across hosts and interactions}

\item{taxon_idx}{optional logratio coordinate to render correlations against; if NULL, render all pairwise correlations}

\item{show_plot}{optional flag to show() plot in addition to rendering it to a file}

\item{return_matrix}{optional flag to return host x interaction correlation value matrix}
}
\value{
NULL or heatmap matrix
}
\description{
Plot heatmap over all pairwise (MAP) correlations between microbes at designated taxonomic level
}
\examples{
Sigmas <- load_MAP_estimates(tax_level="ASV", logratio="clr")
plot_interaction_heatmap(tax_level="ASV", logratio="clr", Sigmas=Sigmas)
}
