% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{plot_ordination}
\alias{plot_ordination}
\title{Plot first several principle coordinates of ordination}
\usage{
plot_ordination(tax_level = "genus", which_measure = "Sigma",
  annotation = "host", MAP = FALSE, show_plot = FALSE)
}
\arguments{
\item{tax_level}{taxonomic level at which to agglomerate data}

\item{which_measure}{estimated object to visualize, either Lambda or Sigma}

\item{annotation}{label to assign (e.g. individual)}

\item{MAP}{use MAP estimate model output instead of full posterior output}

\item{show_plot}{show() plot of first 2 principle coordinates (in addition to rendering first 4 PCoA to files)}
}
\description{
Plot first several principle coordinates of ordination
}
\examples{
plot_ordination(tax_level="genus", which_measure="Sigma", annotation="host", MAP=FALSE)
}
