% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analyses.R
\name{plot_posterior_predictive}
\alias{plot_posterior_predictive}
\title{Plot posterior predictive intervals of Eta (denoised ALR abundances) over observed time series for designated host}
\usage{
plot_posterior_predictive(
  host,
  tax_level = "genus",
  predict_coords = NULL,
  show_plot = FALSE
)
}
\arguments{
\item{host}{host short name (e.g. ACA)}

\item{tax_level}{taxonomic level at which to agglomerate data}

\item{predict_coords}{ALR coordinates to visualize prediction intervals over}

\item{show_plot}{show() plot of first 2 principle coordinates (in addition to rendering first 4 PCoA to files)}
}
\description{
Plot posterior predictive intervals of Eta (denoised ALR abundances) over observed time series for designated host
}
\details{
Output are png files of predicted series intervals.
}
\examples{
plot_posterior_predictive(host="ZIB", tax_level="genus", predict_coords=c(1,2,3))
}
